Extract IntegrationFrame from IntegrationDialog
